+++ app/src/main/java/com/ovidium/comoriod/MainActivity.kt
@@ -110,7 +110,7 @@ fun BottomBarMain(
         }
 
         composable(Screens.Favorites.route) {
-            FavoritesScreen(favoritesModel)
+            FavoritesScreen(favoritesModel, scaffoldState)
         }
 
         composable(
